Update kernel-5.4 v00.04.16
Update kernel-5.15 v00.05.08
Update u-boot-2019.04 v00.04.17
Update aspeed_app v00.01.11
Update openssl v00.01.02
Support kernel-6.6 v00.06.00
Support u-boot-2023.10	v00.05.00

Merge branch 'aspeed-dev' into aspeed-master

+++ meta-aspeed-sdk/classes/uboot-sign-aspeed.bbclass
@@ -98,6 +98,9 @@ UBOOT_FIT_ARM_TRUSTED_FIRMWARE_A_IMAGE ?= "bl31.bin"
 UBOOT_FIT_OPTEE_OS ?= "0"
 UBOOT_FIT_OPTEE_OS_IMAGE ?= "tee-raw.bin"
 
+UBOOT_FIT_UBOOT_LOADADDRESS ?= "${UBOOT_LOADADDRESS}"
+UBOOT_FIT_UBOOT_ENTRYPOINT ?= "${UBOOT_ENTRYPOINT}"
+
 python() {
     # We need u-boot-tools-native if we're creating a U-Boot fitImage
     sign = d.getVar('UBOOT_SIGN_ENABLE') == '1'
@@ -272,8 +275,8 @@ uboot_fitimage_assemble() {
             os = "u-boot";
             arch = "${UBOOT_ARCH}";
             compression = "none";
-            load = <${UBOOT_LOADADDRESS}>;
-            entry = <${UBOOT_ENTRYPOINT}>;
+            load = <${UBOOT_FIT_UBOOT_LOADADDRESS}>;
+            entry = <${UBOOT_FIT_UBOOT_ENTRYPOINT}>;
 EOF
 
 	if [ "${SPL_SIGN_ENABLE}" = "1" ] ; then
@@ -402,7 +405,9 @@ EOF
 			${SPL_MKIMAGE_SIGN_ARGS}
 	fi
 
-	cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	if [ -f "${SPL_DIR}/${SPL_NODTB_BINARY}" ]; then
+		cp ${SPL_DIR}/${SPL_DTB_BINARY} ${SPL_DIR}/${SPL_DTB_SIGNED}
+	fi
 }
 
 uboot_assemble_fitimage_helper() {
